tftp: fix formatting warnings/errors

+++ examples/gnrc_tftp/tftp_client.c
@@ -59,7 +59,7 @@ static bool _tftp_client_start_cb(tftp_action_t action, tftp_mode_t mode,
     }
 
     /* display the action being performed */
-    printf("tftp_client: %s %s %s:%u\n", str_mode, str_action, file_name, *len);
+    printf("tftp_client: %s %s %s:%lu\n", str_mode, str_action, file_name, (unsigned long)*len);
 
     /* return the length of the text, if this is an read action */
     if (action == TFTP_READ) {
@@ -92,7 +92,7 @@ static int _tftp_client_data_cb(uint32_t offset, void *data, size_t data_len)
     }
     else {
         /* we received a data block which we output to the console */
-        printf("\n -- CLIENT DATA --\n%.*s\n -- CLIENT DATA --\n", data_len, c);
+        printf("\n -- CLIENT DATA --\n%.*s\n -- CLIENT DATA --\n", (int)data_len, c);
     }
 
     /* return the length of the data block */
